Rename xmlCreator.js to libvirt-xml-create.js to be consistent
