+++ src/libvirtApi/storagePool.js
@@ -25,7 +25,7 @@ import cockpit from 'cockpit';
 import store from '../store.js';
 
 import { updateOrAddStoragePool } from '../actions/store-actions.js';
-import { getPoolXML } from '../xmlCreator.js';
+import { getPoolXML } from '../libvirt-xml-create.js';
 import { parsePoolCapabilities, parseStoragePoolDumpxml } from '../libvirt-xml-parse.js';
 import { storageVolumeGetAll } from './storageVolume.js';
 import { call, Enum, timeout } from './helpers.js';
